slapd: Add initial configuration.
This configuration is a bit of mix of configurations. It is based on the
slapd.conf from drsnuggles, but the settings match those generated by
debconf. However, debconf generated a slapd.d ldif storage, but since
that is hard to version with git, we stick to the old-style slapd.conf
for now.

Relative to the default configuration, the following things were set:
- Root dn was set to dc=tika,dc=stderr,dc=nl
- A reduced qmail schema was added
- A custom schema (as used on drsnuggles right now) was added
- A few indices were added, to work with the above schemas

Note that the config file also references a bunch of standard schemas,
installed by the Debian package, but those are not added to git since
they're unmodified. Also, the ldapns schema was not included (on
drsnuggles it is used for the hostObject objectclass), since we'll
probably be using a custom objectclass for that.