exim: Add initial configuration.
This configuration is a complete rewrite of the default Debian
configuration, based on the configuration from drsnuggles. This handles
outgoing mail for the host and containers and forwarding and delivery
based on LDAP info.

Spamchecking is configured but still disabled (spamassasin is not set up
yet), authenticated SMTP and SPF/DKIM/DMARC still need to be
investigated and/or configured.