fail2ban: Setup ssh and dovecot filters
This lets both filters just read from rsyslog's all.log, which collects
data from all containers and all services. Their regexes should pick out
the relevant log entries.

Note that this protects only the SSH service running on the login jail,
since only port 22 is blocked (host runs on 222) and blocking rules are
added to the FORWARD chain (host would need the INPUT chain blocked).

This could be fixed, but it requires duplicating the ssh filter
configuration. Since the host runs its SSH on a non-standard port and
only has a handful of users (with strong or disabled passwords), this
shouldn't be a problem.