nslcd: maroesjaSystemIdentity now represents a user and group
Previously, a lot of users in the LDAP did not have a named primary
group: They just used their uniqueIdentifier / uidNumber as their
gidNumber as well, without referring to an actual group. Now, the LDAP
object for each user also has a cn and represents a group as well as a
user, making the primary group for all users have the same name as the
user itself.