slapd: Give maroesjaSystemIdentity the "cn" attribute
This causes the maroesjaSystemIdentity to represent both a user and its
own primary group. Since it's not a maroesjaGroup, it cannot contain any
other members, but this helps to give each user a meaningful primary
group (previously a lot of users just had an unnamed primary group with
the same gidNumber as their own uidNumber).