spamassassin: Increase --max-children to 10
This used to be 5, but occasionally a spike in incoming mails would run
out of room to spawn new spamd instances. This wasn't a big problem by
itself, since up to 128 new connections are queued, but allowing a few
more instances might make things more smooth (though these spamd
instances fight for the same CPU resources, so it's probably only a
small gain).

Furthermore, increasing max-children prevents warnings about it from
showing up in the logs all the time, but still allow the warning to
attract attention when there is a really big load spike or other
problem.from showing up in the logs all the time, but still allow the
warning to attract attention when there is a really big load spike or
other problem.