exim: Fix ldap_group_member router
Due to an upgrade of exim, the way multiple "member" attributes were
returned has changed and this accounts for that.

Previously, these were joined using ", " and any commas in the member
dns were untouched, requiring (and allowing) a hack replacing ", " with
newlines and then splitting on newlines to access the individual member
dns.

In more recent versions, the multiple dns are separated just by ",", but
any commas inside the dns are doubled. This allows using exim's
alternate separator syntax {<, ...} directly to split on commas, while
at the same time undoubling the commas.