Add systemd generator to support bindmount options
Normally, bindmounts initially copy their options from the original
filesystem. To change options just on the bindmount, the bind must be
remounted with the required options.

Classic fstab mounting using mount -a supports this by having two
entries for the same mount point, the second with the remount and
required options (this only works with mount -a, which just executes all
mounts in order, but when specifying the mount point explicitely only
the first matching entry is used).

When using systemd, multiple entries for the same mount point are not
supported at all. To still support options, this commit adds a systemd
generator, that scans fstab (just like the builtin
systemd-fstab-generator already does) and generates unit files to
execute remounts automatically for any fstab entry that specifies the
bind option, along with other options.

The generator script is stored in /lib/systemd/system-generators.
Ideally it would be in /etc/systemd/system-generators, but this is only
supported in systemd 219 and onwards.

This generator script is taken from
https://blog.iwakd.de/systemd-fstab-and-bind-mounts-with-options