apt: Make stretch the new default release
This shuffles around sources.list to become consistent with the sources
of the containers, containing all current releases. The source lines are
also removed. The vuurmuur source was updated from wheezy to sid, since
none of the intermediate releases seem to be available.

Additionally, the pinning configuration is changed to make "stretch" the
default (instead of "stable"), to ensure that upgrading to the next
release requires explicit action in the future.