exim: Scan messages up to 100k at the gate
Previously, only messages up to 20k where scanned at SMTP time, and up
to 100k during delivery. This caused a large number of messages to be
accepted, and later marked as spam. For messages that were externally
forwarded, this caused a lot of outgoing spam.

This should reduce outgoing spam on forwarding addresses, improving our
reputation.