certbot: reload nginx on renew
This modifies the systemd certbot service file to use a config file. The
config file then defines a post-hook that reloads nginx.

This is an older change, that was previously implemented by modifying
the cron file for certbot, but this was never committed. With the
stretch upgrade, a systemd timer is used when available.