lxc: Enable hairpin mode for all container interfaces
This allows traffic to be routed back to a container, e.g. when the web
container tries to connect to itself through the external IP address.
This probably worked before, so it likely broke with a kernel upgrade,
or perhaps the installation of the docker.io package and went unnoticed
for a while.