slapd: Allow reading of passwordLastChange
Previously, this attribute was configured along with passwordLastChange
so users could modify their own (which is good), but also that
unauthenticated connections cannot read it. The latter is good for
passwords, but it seems that nslcd reads passwordLastChange without
authenticating, so that would not work.

By configuring both attributes separately, the passwordLastChange can
again be made public.