lxc: Update configs to buster
Most conversion happened using lxc-update-config (after fixing
https://github.com/lxc/lxc/issues/3369 manually).

On top of that, lxc.hook.version was set. Without this, info is passed
to hooks as arguments, but the argument order had changed, so the hook
needed to be adapted. With lxc.hook.version set to 1, info is passed in
environment variables, which is more future-proof.