systemd: Remove local-bind-remount-generator
This was added as a custom generator to support options on bind mounts,
but it seems this now works without the generator as well (at least for
the ro options used in fstab). Not sure if systemd became smarter or
bindmounts now support options without remount, but it works, so this
can be removed.